+++ core/src/main/kotlin/riven/core/models/workflow/engine/state/StepOutput.kt
@@ -1,4 +1,4 @@
-package riven.core.models.workflow.engine.datastore
+package riven.core.models.workflow.engine.state
 
 import riven.core.enums.workflow.WorkflowStatus
 import java.time.Instant
