+++ core/src/main/kotlin/riven/core/models/workflow/engine/state/WorkflowState.kt
@@ -1,4 +1,4 @@
-package riven.core.models.workflow.engine.coordinator
+package riven.core.models.workflow.engine.state
 
 import java.util.*
