+++ core/src/main/kotlin/riven/core/models/workflow/node/config/WorkflowNodeConfig.kt
@@ -5,8 +5,8 @@ import io.swagger.v3.oas.annotations.media.Schema
 import riven.core.deserializer.WorkflowNodeConfigDeserializer
 import riven.core.enums.workflow.WorkflowNodeType
 import riven.core.models.common.json.JsonObject
-import riven.core.models.workflow.engine.datastore.NodeOutput
-import riven.core.models.workflow.engine.datastore.WorkflowDataStore
+import riven.core.models.workflow.engine.state.NodeOutput
+import riven.core.models.workflow.engine.state.WorkflowDataStore
 import riven.core.models.workflow.node.NodeServiceProvider
 import riven.core.models.workflow.node.config.validation.ConfigValidationResult
