+++ core/src/main/kotlin/riven/core/models/workflow/node/config/actions/WorkflowQueryEntityActionConfig.kt
@@ -10,9 +10,8 @@ import io.swagger.v3.oas.annotations.media.Schema
 import riven.core.enums.workflow.WorkflowActionType
 import riven.core.enums.workflow.WorkflowNodeConfigFieldType
 import riven.core.enums.workflow.WorkflowNodeType
-import riven.core.models.workflow.engine.datastore.NodeOutput
-import riven.core.models.workflow.engine.datastore.QueryEntityOutput
-import riven.core.models.workflow.engine.datastore.WorkflowDataStore
+import riven.core.models.workflow.engine.state.NodeOutput
+import riven.core.models.workflow.engine.state.WorkflowDataStore
 import riven.core.models.workflow.node.NodeServiceProvider
 import riven.core.models.workflow.node.config.WorkflowActionConfig
 import riven.core.models.workflow.node.config.WorkflowNodeConfigField
@@ -351,7 +350,7 @@ data class WorkflowQueryEntityActionConfig(
 
         throw NotImplementedError(
             "QUERY_ENTITY execution requires EntityQueryService implementation. " +
-                "Filter: ${query.filter}, Pagination: $pagination"
+                    "Filter: ${query.filter}, Pagination: $pagination"
         )
     }
 }
