+++ core/src/main/kotlin/riven/core/models/workflow/node/config/trigger/WorkflowWebhookTriggerConfig.kt
@@ -10,8 +10,8 @@ import riven.core.enums.workflow.WorkflowTriggerType
 import riven.core.models.common.json.JsonObject
 import riven.core.models.common.http.Signature
 import riven.core.models.common.validation.Schema
-import riven.core.models.workflow.engine.datastore.NodeOutput
-import riven.core.models.workflow.engine.datastore.WorkflowDataStore
+import riven.core.models.workflow.engine.state.NodeOutput
+import riven.core.models.workflow.engine.state.WorkflowDataStore
 import riven.core.models.workflow.node.NodeServiceProvider
 import riven.core.models.workflow.node.config.WorkflowNodeConfigField
 import riven.core.models.workflow.node.config.WorkflowTriggerConfig
