+++ core/src/test/kotlin/riven/core/models/workflow/engine/state/NodeOutputTest.kt
@@ -1,4 +1,4 @@
-package riven.core.models.workflow.engine.datastore
+package riven.core.models.workflow.engine.state
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
