+++ core/src/test/kotlin/riven/core/service/workflow/WorkflowExecutionEndToEndIntegrationTest.kt
@@ -11,12 +11,12 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import riven.core.entity.workflow.WorkflowEdgeEntity
 import riven.core.enums.workflow.WorkflowStatus
-import riven.core.models.workflow.engine.coordinator.WorkflowExecutionPhase
-import riven.core.models.workflow.engine.coordinator.WorkflowState
-import riven.core.models.workflow.engine.datastore.GenericMapOutput
-import riven.core.models.workflow.engine.datastore.StepOutput
-import riven.core.models.workflow.engine.datastore.WorkflowDataStore
-import riven.core.models.workflow.engine.datastore.WorkflowMetadata
+import riven.core.models.workflow.engine.state.WorkflowExecutionPhase
+import riven.core.models.workflow.engine.state.WorkflowState
+import riven.core.models.workflow.engine.state.GenericMapOutput
+import riven.core.models.workflow.engine.state.StepOutput
+import riven.core.models.workflow.engine.state.WorkflowDataStore
+import riven.core.models.workflow.engine.state.WorkflowMetadata
 import riven.core.models.workflow.node.WorkflowNode
 import riven.core.models.workflow.node.config.actions.WorkflowCreateEntityActionConfig
 import riven.core.service.workflow.engine.coordinator.WorkflowGraphCoordinationService
