+++ core/src/test/kotlin/riven/core/service/workflow/state/WorkflowNodeInputResolverServiceTest.kt
@@ -11,9 +11,9 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean
 import riven.core.configuration.auth.WorkspaceSecurity
 import riven.core.enums.util.OperationType
 import riven.core.enums.workflow.WorkflowStatus
-import riven.core.models.workflow.engine.coordinator.WorkflowExecutionPhase
-import riven.core.models.workflow.engine.coordinator.WorkflowState
-import riven.core.models.workflow.engine.datastore.*
+import riven.core.models.workflow.engine.state.WorkflowExecutionPhase
+import riven.core.models.workflow.engine.state.WorkflowState
+import riven.core.models.workflow.engine.state.*
 import riven.core.service.auth.AuthTokenService
 import java.time.Instant
 import java.util.*
